Revert "feat: moving cache sync timeout to controller level (#1576)"
This reverts commit eafe86952fcef8f0eede0d9bb2e850098da7f9db.
