+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/api/reconciler/ControllerConfiguration.java
@@ -118,7 +118,4 @@ public @interface ControllerConfiguration {
    *         accessible no-arg constructor.
    */
   Class<? extends RateLimiter> rateLimiter() default LinearRateLimiter.class;
-
-  CacheSyncTimeout cacheSyncTimeout() default @CacheSyncTimeout(
-      timeout = CacheSyncTimeout.DEFAULT_TIMEOUT);
 }
