+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/event/EventSourceManager.java
@@ -48,7 +48,6 @@ public class EventSourceManager<P extends HasMetadata>
 
   public void postProcessDefaultEventSourcesAfterProcessorInitializer() {
     eventSources.controllerResourceEventSource().setEventHandler(controller.getEventProcessor());
-
     eventSources.retryEventSource().setEventHandler(controller.getEventProcessor());
   }
 
@@ -120,7 +119,6 @@ public class EventSourceManager<P extends HasMetadata>
     registerEventSource(null, eventSource);
   }
 
-  @SuppressWarnings("unchecked")
   public final synchronized void registerEventSource(String name, EventSource eventSource)
       throws OperatorException {
     Objects.requireNonNull(eventSource, "EventSource must not be null");
