+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/event/source/informer/InformerManager.java
@@ -97,8 +97,7 @@ public class InformerManager<T extends HasMetadata, C extends ResourceConfigurat
   private InformerWrapper<T> createEventSource(
       FilterWatchListDeletable<T, KubernetesResourceList<T>, Resource<T>> filteredBySelectorClient,
       ResourceEventHandler<T> eventHandler, String key) {
-    var source = new InformerWrapper<>(filteredBySelectorClient.runnableInformer(0),
-        configuration);
+    var source = new InformerWrapper<>(filteredBySelectorClient.runnableInformer(0));
     source.addEventHandler(eventHandler);
     sources.put(key, source);
     return source;
