+++ operator-framework-core/src/test/java/io/javaoperatorsdk/operator/ControllerManagerTest.java
@@ -58,7 +58,7 @@ public class ControllerManagerTest {
     public TestControllerConfiguration(Reconciler<R> controller, Class<R> crClass) {
       super(null, getControllerName(controller),
           CustomResource.getCRDName(crClass), null, false, null, null, null, null, crClass,
-          null, null, null, null, null, null, null);
+          null, null, null, null, null, null);
       this.controller = controller;
     }
