+++ operator-framework-core/src/test/java/io/javaoperatorsdk/operator/processing/event/source/ResourceEventFilterTest.java
@@ -145,7 +145,7 @@ class ResourceEventFilterTest {
           eventFilter,
           customResourceClass,
           null,
-          null, null, null, null, null, null);
+          null, null, null, null, null);
     }
   }
