+++ operator-framework/src/test/java/io/javaoperatorsdk/operator/InformerRelatedBehaviorITS.java
@@ -205,15 +205,12 @@ class InformerRelatedBehaviorITS {
     Operator operator = new Operator(clientUsingServiceAccount(),
         co -> {
           co.withStopOnInformerErrorDuringStartup(stopOnInformerErrorDuringStartup);
+          co.withCacheSyncTimeout(Duration.ofMillis(3000));
           if (addStopHandler) {
-            co.withInformerStoppedHandler((informer, ex) -> {
-              replacementStopHandlerCalled = true;
-            });
+            co.withInformerStoppedHandler((informer, ex) -> replacementStopHandlerCalled = true);
           }
         });
-    operator.register(reconciler, co -> {
-      co.withCacheSyncTimeout(Duration.ofMillis(3000));
-    });
+    operator.register(reconciler);
     operator.installShutdownHook();
     operator.start();
     return operator;
